created checkCredentials.js module which exports isLoggedIn and isAdmin in the api routes
