+++ server/api/orders.js
@@ -8,28 +8,10 @@ const {
   Order,
   OrderItem
 } = require('../db/models')
+const {isLoggedIn, isAdmin} = require('./checkCredentials')
 const Sequelize = require('sequelize')
 const Op = Sequelize.Op
 module.exports = router
-
-const isLoggedIn = (req, res, next) => {
-  if (req.user) next()
-  else {
-    const err = new Error('Must loggin to do things')
-    res.status(401)
-    next(err)
-  }
-}
-
-const isAdmin = (req, res, next) => {
-  if (req.user.userType === 'admin') {
-    next()
-  } else {
-    const err = new Error('Must be an admin')
-    res.status(401)
-    next(err)
-  }
-}
 
 //UPDATE ORDER
 router.put('/:id', async (req, res, next) => {
